Redesign: contacts_book responsible for editing contacts in db.
